Decouple Flask `app` from `gevent` monkeypatching
